refactor: MapStruct 사용에 따른 기존 컨버터 클래스 삭제
